+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - compatible with old vscode version.
 - Fix code len detect comment as sql.
 - Add context menu to result view.
+- Add more chinese content.
 
 # 3.2.2 2020/1/25